guroomed-master/requirements.txt
Django==2.0.2
gunicorn==19.7.1
psycopg2==2.7.1
django-embed-video==1.1.2
django-role-permissions==2.0.1
PyVimeo==0.3.10
social-auth-app-django==2.1.0
elasticsearch-dsl==5.3.0
elasticsearch==5.4.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
requests==2.9.1
paypalrestsdk==1.13.1
simplejson==3.13.2
PyJWT==1.5.3
django-paypal==0.4.1